Engadida optimizacion Middlewares + Optimizacions

+++ 5_tpIntegradorBack/index.js
@@ -10,6 +10,8 @@ const PORT = environments.port;
 import connection from "./src/api/database/db.js"; // Importamos la conexion a la BBDD para poder enviarle sentencias SQL
 import cors from "cors";
 
+import { loggerUrl, validateId } from "./src/api/middlewares/middlewares.js";
+
 
 /*====================
     Middlewares
@@ -17,13 +19,7 @@ import cors from "cors";
 app.use(cors()); //Middleware CORS basico que permite todas las solicitudes
 app.use(express.json()); // Middleware que transforma el JSON de las peticiones POST y PUT a objetos JS
 
-// Middleware logger para mostrar por consola todas las peticiones a nuestro servidor
-app.use((req, res, next) => {
-    console.log(`[${new Date().toLocaleString()}] ${req.method} ${req.url}`);
-   
-    // Con next continuamos al siguiente middleware o a la respuesta
-    next();
-});
+app.use(loggerUrl);
 
 
 /*==================
@@ -38,16 +34,19 @@ app.get("/", (req, res) => {
 app.get("/products", async (req, res) => {
 
     try {
+        // Optimizacion 1: Seleccionar solamente los campos necesarios, evitar SELECT *
+        // La idea es devolver solo las columnas que necesita el front: - datos transferidos, - carga de red, + seguridad
         const sql = "SELECT * FROM productos";
     
         // Con rows extraemos exclusivamente los datos que solicitamos en la consulta
         const [rows] = await connection.query(sql);
 
         // Comprobamos que se reciban correctamente los productos
-        // console.log(rows);
+        //console.log(rows);
         
         res.status(200).json({
-            payload: rows
+            payload: rows,
+            message: rows.length === 0 ? "No se encontraron productos" : "Productos encontrados"
         });
         
     
@@ -62,20 +61,44 @@ app.get("/products", async (req, res) => {
 
 
 // Get product by id -> Consultar producto por su id
-app.get("/products/:id", async (req, res) => {
+app.get("/products/:id", validateId, async (req, res) => {
 
     try {
         // let id = req.params.id;
         let { id } = req.params; // Aca extraemos el valor "2" de localhost:3000/products/2
+
+        /* Logica exportada al middleware validateId
+        // Optimizacion 1: Validacion de parametros antes de acceder a la BBDD para evitar hacer una consulta donde el parametro id no sea valido
+        if(!id || isNaN(Number(id))) {
+            return res.status(400).json({
+                message: "El id del producto debe ser un numero valido"
+            })
+        }
+        */
 
         // Gracias al uso de los placeholders -> ? evitamos ataques de inyeccion SQL
         //let sql = `SELECT * FROM productos WHERE productos.id = ${id}`; // Opcion 1. Consulta no segura
         let sql = "SELECT * FROM productos WHERE productos.id = ?"; // Opcion 2, sentencia mas segura
 
         //let [rows] = await connection.query(sql); // Aca introducimos la consulta 1 no segura
-        let [rows] = await connection.query(sql, [id]); // Este id reemplazara el placeholder ?
 
+        // Optimizacion 2: Limitar los resultados de la consulta
+
+        
+        let [rows] = await connection.query(sql, [id]); // Este id reemplazara el placeholder ?
         //console.log(rows);
+        
+        // Optimizacion 3: Comprobamos que exista el producto con ese id
+        if(rows.length === 0) {
+            // Este console.log es desde la consola del servidor
+            console.log(`Error! No se encontro producto con id ${id}`);
+
+            // Esta respuesta se la brindamos al usuario y puede elegir verla por consola o por pantalla
+            return res.status(404).json({
+                message: `No se encontro producto con id ${id}`
+            });
+        }
+
  
         res.status(200).json({
             payload: rows,
@@ -125,10 +148,6 @@ app.get("/products/:id", async (req, res) => {
 })
 
 
-/* TO DO: Optimizar:
-    - GET: Seleccionar solamente los campos necesarios y devolver tambien un message junto al payload
-    - GET by id: Meter optimizaciones
-*/
 
 
 // POST -> Crear un nuevo producto
@@ -143,16 +162,24 @@ app.post("/products", async (req, res) => {
 
         // Gracias al destructuring, recogemos estos datos del body
         let { image, name, price, type } = req.body;
-
         console.log(req.body);
+
+        // Optimizacion 1: Validacion de datos de entrada
+        if(!image || !name || !price || !type) {
+            return res.status(400).json({
+                message: "Datos invalidos, asegurate de enviar todos los campos"
+            });
+        }
 
         let sql = `INSERT INTO productos (imagen, nombre, precio, tipo) VALUES (?, ?, ?, ?)`;
 
-        let [rows] = await connection.query(sql, [image, name, price, type]);
+        let [result] = await connection.query(sql, [image, name, price, type]);
+        console.log(result);
 
         // Codigo de estado 201 -> Created
         res.status(201).json({
-            message: "Producto creado con exito"
+            message: "Producto creado con exito",
+            productId: result.insertId
         });
 
     } catch(error) {
@@ -171,6 +198,13 @@ app.put("/products", async (req, res) => {
     try {
         let { id, name, image, type, price, active } = req.body;
 
+        // Optimizacion 1: Validacion basica de datos recibidos
+        if(!id || !name || !image || !type || !price || !active) {
+            return res.status(400).json({
+                message: "Faltan campos requeridos"
+            });
+        }
+
         let sql = `
             UPDATE productos
             SET nombre = ?, imagen = ?, tipo = ?, precio = ?, activo = ?
@@ -178,8 +212,14 @@ app.put("/products", async (req, res) => {
         `;
 
         let [result] = await connection.query(sql, [name, image, type, price, active, id]);
-
         console.log(result);
+
+        // Optimizacion 2: Testeamos que se actualizara, esto lo sabemos gracias a affectedRows que devuelve result
+        if(result.affectedRows === 0) { // No se actualizo nada
+            return res.status(400).json({
+                message: "No se actualizo el producto"
+            });
+        }
 
         res.status(200).json({
             message: `Producto con id ${id} actualizado correctamente`
@@ -196,9 +236,18 @@ app.put("/products", async (req, res) => {
 
 
 // DELETE -> Eliminar un producto por su id
-app.delete("/products/:id", async (req, res) => {
+app.delete("/products/:id", validateId, async (req, res) => {
     try {
         let { id } = req.params;
+
+        /* Logica exportada al middleware validateId
+        // Optimizacion 1: Validacion de parametros antes de acceder a la BBDD para evitar hacer una consulta donde el parametro id no sea valido
+        if(!id || isNaN(Number(id))) {
+            return res.status(400).json({
+                message: "El id del producto debe ser un numero valido"
+            })
+        }
+        */
 
         // Opcion 1: Borrado normal
         let sql = "DELETE FROM productos WHERE id = ?";
@@ -208,6 +257,14 @@ app.delete("/products/:id", async (req, res) => {
 
         let [result] = await connection.query(sql, [id]);
         console.log(result);
+
+        // Optimizacion 2: Testeamos que se borro, esto lo sabemos gracias a affectedRows que devuelve result
+        if(result.affectedRows === 0) { // No se borro nada
+            return res.status(400).json({
+                message: "No se eliminó el producto"
+            });
+        }
+        
 
         return res.status(200).json({
             message: `Producto con id ${id} eliminado correctamente`
@@ -223,8 +280,6 @@ app.delete("/products/:id", async (req, res) => {
     }
 });
 
-
-// TO DO, optimizar endpoints y mostrar mensaje de error
 
 
 app.listen(PORT, () => {
